Implement: object-version-control
Warning: This is very experimental and not ready for production use.

This is a object version control system. It is like version control systems like Git.
But it is designed to work with Plain JavaScript Objects. (POJO)

This packages uses Automerge and jsondiffpatch. But it is used in a way to merge objects.
You can choose or implement other merge strategies for more complex use cases. (domain specific merge strategies)

```js
const ovc = ObjectVersionControl.create()

// Commit a snapshot
const commitHash1 = ovc.commit({ key: 'value' })
```