Fix ArrayBufferLike in tests
The following error occurred during type checking.

```
Error: packages/cipher/src/xorShuffle.spec.ts(19,7): error TS2322: Type 'Uint8Array<ArrayBufferLike>' is not assignable to type 'Uint8Array<ArrayBuffer>'.
  Type 'ArrayBufferLike' is not assignable to type 'ArrayBuffer'.
    Type 'SharedArrayBuffer' is not assignable to type 'ArrayBuffer'.
      Types of property '[Symbol.toStringTag]' are incompatible.
        Type '"SharedArrayBuffer"' is not assignable to type '"ArrayBuffer"'.
```

This is due to a compatibility issue between `ArrayBufferLike` and `ArrayBuffer`. It is a false positive.
This commit resolves the issue by adding type assertions to the variables in the test code.